requirements.txt
# FILE: requirements.txt
anthropic>=0.25.0
google-generativeai>=0.3.0
openai>=1.0.0
psycopg2-binary
boto3
kafka-python
qdrant-client>=1.6.0
numpy>=1.24.0
pandas>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
python-multipart>=0.0.6
websockets>=11.0.0
PyPDF2>=3.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
loguru>=0.7.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0